chat/pyproject.toml
[tool.poetry]
name = "agent"
version = "0.2.0"
description = "一个简单的cli gui和webAI Agent智能体程序"
authors = ["'FreeRUOK' <'2651688427@qq.com'>"]
license = "MIT"
readme = "README.md"
packages = [{include = "chat"}]

[tool.poetry.dependencies]
python = "^3.11"
typer = "^0.15.1"
ollama = "^0.6.1"
pydub = "^0.25.1"
simpleaudio = "^1.0.4"
typing-extensions = "^4.12.2"
pyperclip = "^1.9.0"
prompt-toolkit = "^3.0.50"
edge-tts = "^7.0.0"
pyyaml = "^6.0.2"
chardet = "^5.2.0"
wxpython = "^4.2.2"
loguru = "^0.7.3"
flask = "^3.1.0"
flask-socketio = "^5.5.1"
pillow = "^11.1.0"
pyautogui = "^0.9.54"
pygetwindow = "^0.0.9"
opencv-python = "^4.11.0.86"
vosk = "^0.3.45"
sounddevice = "^0.5.5"
pvporcupine = "^4.0.2"
baidusearch = "^1.0.3"
requests = "^2.32.5"
ddgs = "^9.10.0"
pydantic = "^2.12.5"
openai = "^2.21.0"
dotenv = "^0.9.9"
chinese-calendar = "^1.11.0"
lunar-python = "^1.4.8"
pint = "^0.25.2"
simpleeval = "^1.0.3"
diskcache = "^5.6.3"


[tool.poetry.group.dev.dependencies]
ruff = "^0.9.4"
mypy = "^1.15.0"
types-pyyaml = "^6.0.12.20241230"
types-flask-socketio = "^5.5.0.20241227"
nuitka = "^2.6.8"
types-pyautogui = "^0.9.3.20241230"
bandit = "^1.9.3"
types-requests = "^2.32.4.20260107"
pytest = "^9.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"
priority = "primary"


[[tool.poetry.source]]
name = "tencent"
url = "https://mirrors.cloud.tencent.com/pypi/simple/"
priority = "primary"


[[tool.poetry.source]]
name = "tsinghua"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
priority = "primary"

[tool.mypy]
check_untyped_defs = true

[tool.poetry.scripts]
chat = "chat.__main__:app"
